[java] Code cleanup using Eclipse source cleanup

+++ main/src/main/java/mockit/asm/constantPool/ConstantPoolGeneration.java
@@ -442,7 +442,7 @@ public final class ConstantPoolGeneration {
         }
 
         if (cst instanceof Character) {
-            return newInteger((int) (Character) cst);
+            return newInteger((Character) cst);
         }
 
         if (cst instanceof Boolean) {
